Using DISTINCT instead of GROUP BY breaks distance
Distance and area use aggregate calculations that require a group. We want to use DISTINCT so count returns an integer instead of a hash, but can't in all situations. However, in the situations where we have to use GROUP BY, we won't want to be counting, so it's all good.