Improve total area calculation speed
The previous implementation was being extremely inefficient with how it flattened geometry before intersecting. It performed a intersection join, and then flattened all geometry for the current record (A), and again for the records overlapping it (B). But since the join results is the cartesian cross product, there could be thousands of duplicate geometry being flattened.

Instead, we improve the algorithm by first finding the features that overlap, then creating subqueries where we flatten the geometry for  A and B separately, then perform the intersection query on the flattened results. This avoids unnecessarily unioning features with themselves, and speeds up the calculation over 10x.