Add methods to maintain a spatial cache
update_proximity can be used in place of cache_proximity in situations where only some of the records will change between each invocation. It operates per-record and ensures the given caches are kept up to date. It skips caches that are not stale, and creates any caches that are missing.