Fix import validation
When assigning a collection, it raises an exception if any of the records can’t be saved. This means we didn’t reach the optional validation code in `validate_features!` before exploding. This commit partitions the records based on which saved, and then prepares an error message if we are not skipping invalid.