Optimize intersection area lookup
If there is a spatial proximity record between this record and other, whose intersection area is equal (or greater) to the total area of this record, we can assume the record is fully intersected by `other`, and we just return the area of this record.

This avoids computing the union of all the features in `other`, and intersecting with that union, and calculating the area, any of which may be very expensive calculations depending on polygon complexity.