Fix bug where feature area wasn’t updated after save
Because the update was triggered `after_save`, it competed with the `update_features!` callback. Since it was declared last, it was triggered first, meaning we checked if the features has been updated before they were updated. This fixes the issue by explicitly updating during `update_features!`, as well as after save. This ensures that if features are set manually, or via `update_features!`, the area will be recalculated.