Refactored fusion tables integration to support colour
We need to be able to store the layer colour in fusion tables because it turns out Google only allows you to customize one layer using the Google Maps API. Thus, all layer colour need to be customized when the layer is stored in the Fusion Table instead of customized in the browser using JS.

This commit ensures the stored colour is coerced to a compatible `#RRGGBBAA` format so we can use it to colour each polygon. It also refactors the callback for updating the fusion table so multiple events can `expire` the table, and a single check can see if it is expired and update it when all of the expiration events have completed. This allows the table to update once, instead of multiple times unintentionally.