Unscope own class on spatial cache lookup
Because we can be within a scope when calling `cached_within_buffer_scope`, i.e. ActiveRecord wraps method calls delegated from a relation to the `klass` in a scope, we need to ensure we unscope when looking up the spatial cache because its `spatial_model` association could inadvertantly be scoped now that we are allowing same-class spatial intersection queries.