Fix bug when intersecting subclasses
We had an issue where the foreign type columns in spatial caching were not following the Rails convention of storing the base class name even when the foreign key was to a subclass. This manifested itself in two places, on the polymorphic foreign key to a spatially cached record, and on the model name we store on the `SpatialCache` `intersection_model_type` column.

To avoid this issue, we ensure we always refer to and store the base class for the polymorphic associations. To fix the `intersection_model_type` we don’t want to coerce it to the base class because that would incorrectly state that we have a `SpatialCache` on the base class and all subclasses instead of only on a subset of the subclasses. Instead, whenever we scope using the `intersection_model_type`, we include the given type and all of its ancestor classes. This is safe because if we have a cache on the ancestor class, it must include the `intersection_model_type` subclass.