Fix bug where Shapefile could unzip the wrong file
In a shapefile there is a .shp.xml file. If this file happened to be before the .shp file in the archive, the importer would return the wrong file name and import would fail.

Instead, we modify the Unzip to accept a regexp as the desired file pattern, allowing us to specify an extension anchored to the end of the filename string.