Fix issue where intersection area could explode
We were having issues where finding the intersection area of certain geometry would return a TopologyException: found non-noded intersection between LINESTRING (1.69029e+06 504904, 1.69029e+06 504907) and LINESTRING (1.69029e+06 504907, 1.69029e+06 504907) at 1690288.5152031549 504906.78335195634.

A first attempt at fixing this using ST_MakeValid did not work. ST_Buffer was found to solve the problem, but required a non-zero buffer. Ultimately I settled on using the existing ST_Union, but moving it before the intersection in order to union all geometry each scope before we intersect the two. The thought process is that this should eliminate any strange self interaction between all the features in either set, essentially “flattening” each before we start finding intersections.

I was unable to replicate this error in a test, even when using the same “geog” data from the error, possibly because the intersecting polygon is different or differences with the SRID.