Pre-aggregate features to optimize query
Spatial joins can be expensive because we end up unioning all the features for each record on-the-fly. This means we’re not able to take advantage of precalculated indices for intersection.

Instead, we maintain an aggregate feature for all the features on a spatial model and use that when joining, avoiding the need to `ST_Union` each spatial model’s features on each query.