Defer generation of aggregate features on import
Since we’re importing many features for the same spatial model during an import, it does not make sense to recalculate the aggregate feature each time we save a `Feature`.

Instead, we defer the aggregate feature refresh until the features have been imported and then run it once for each spatial model imported.