Add support for geoJSON output
This commit adds support for converting features to geoJSON. geoJSON output must be specially formatted if it is to work with Google Maps. Specifically, it must be converted to a `FeatureCollection` and contain `Feature` objects. See https://postgis.net/docs/ST_AsGeoJSON.html for an example. Unfortunately `ST_AsGeoJSON` does not output this format, and we must assemble to boilerplate around that function’s output.

The geojson mime type is also included to make serving geojson simpler.