Fix PostGIS 3.x compatibility
It appears as though PostGIS 2.x would happily ignore empty geometry in `ST_Collect`, or perhaps empty geometry would retain the SRID it was derived from. Whatever the cause, now when we `ST_CollectionExtract`, if there is no geometry, it returns an empty collection whose SRID is 0. This causes an explosion if any of the other extracts returns geometry, because `ST_Collect` doesn’t want mixed SRIDs. To work around this, we simply remove any empty sets from the array before collecting.

Alternatively, we could just not extract, but in 2.x testing it was found to be faster to extract homogenous sets of geometry, union them, and then reassemble through `ST_Collect`. It should not impact performance to perform 3 removals on an array of 3 or fewer elements, so we go with this method instead.