Fixed bug with cached, same-class spatial queries
With heterogenous-only caching, we were guaranteed to be able to determine which column, (a or b) each record would be in in the spatial proximity cache. This is because we compared class names and decided alphabetically. Now that we support homogenous caching too, we are no longer guaranteed to know which column to look in for each record. Therefore, we must now check both combinations record in column a and other record in column b, and record in column b and other record in column a.

Since it no longer matters which column each record appears in, I have removed the code that would make that decision when creating the proximity cache.