Tweak PostGIS 3.x compat fix.
In b3f876150036f647035b9b9b4d587bccb9b953c7, we added a fix for PostGIS 3.x. This fix was still failing if the CollectionExtract returned a Multi* geometry since we weren’t removing empty Multis.

One example of this is when refreshing the aggregate feature for a community.

We now convert all geometries to Multis before removing empties to ensure the correct behaviour.

This is starting to turn into a bit of a Rube Goldberg machine so we should see if there is a simpler way to do this at some future point.