Tweak PostGIS 3.x compat fix.
In d0c2e54c1a07aa935a905a30e5be2dc443eb1a7f, we added a fix for PostGIS 3.x. This fix was still failing if the CollectionExtract returned a Multi* geometry since we weren’t removing empty Multis.

One example of this is when refreshing the aggregate feature for Community 16 in Sto:lo Connect.

We now convert all geometries to Multis before removing empties to ensure the correct behaviour.

This is starting to turn into a bit of a Rube Goldberg machine so we should see if there is a simpler way to do this at some future point.