Fix bug where Feature callbacks fired unnecessarily
In the importer we create the features and then assign them to the spatial model association so they are cached, however, at some point this triggers a save on those records. Since the geog has not been changed, there’s no need to recalculate the derivatives. The same was happening for the aggregate feature refresh callback.