Optimize within buffer calculations
Instead of generating geometry and checking for intersections, we now use the `ST_DWithin` function which has been optimized since earlier versions Postgis. This allows us to perform a within buffer calculation on a geometry collection with 10000 points in less than a second, whereas before the system would run out of memory while it generated geometry for a 10km buffer.