Attempt to clean up aggregate geometry
We were getting a segfault when performing a spatial join between NSTQ’s ReferralSubmission 505 and any other geometry. To combat this, we homogenize the geometry collection in an attempt to avoid whatever code path leads to this segfault. Buffering also worked, but had the effect of breaking tests, probably because even with a buffer of 0, geometry is modified, affecting some of tests where we worked with very small shapes.