Fix bug where `intersects?` would fail when glint indexing
Under Rails 4.2.11.3, asking a record if it `intersects?` another relation it would fail with a "ActiveRecord::ConfigurationError: nil" unless we first unscoped the current class. This should not affect the query since we’re not relying on any scoped queries anyway. However, it is a strange issue that is potentially fixed in later versions of Rails.