Use aggregate features cache key
We can optimize cache key generation by using the aggregate feature. This improves performance in situations where there may be hundreds of thousands of features, and counting becomes slow. Using the aggregate feature allows the complexity of that query to scale with the number of spatial models instead of the number of features on all those spatial models.