Skip introspecting columns when database tables don't exist yet
The `with_stale_spatial_cache` scope is only defined when the model has
a `features_hash` column.  We won't have any tables to inspect when we're
running this code before the database is setup (e.g. CI).

We short-circuit to `false` which temporarily gives us a reduced feature set
but prevents errors.

Fixes #10