Optimize cache keys
We have flip-flopped on this in the past going from two to one to two queries. It is my opinion that for large tables, Postgres seems to perform fastest with two separate queries. The steady state performance of the maximum(id) query is almost instaneous, arriving in the sub-millisecond range. Similarly the count query tops out at about 10ms filtering for AggregateFeatures results on a table with 450k rows. These results are in contract with the best single query performance which was a much slower 130ms.