Tweak JSON import
From what I can tell ogr2ogr can accept an ESRI JSON or GeoJSON file and output OGRGeoJSON, without changing the input args. Therefore the JsonArcGIS importer is not limited to only JSON from ArcGIS, it's a general JSON importer. It does have the ability to import paginated JSON from an ArcGIS server, but it can also import from any source, be it a url or a file path. In fact, ogr2ogr appears to handle KML, and Shapefiles, so we can probably just refer to it as OGR in anticipation of using it as a replacement for nearly all importers.