Expand GeoJSON generation
GeoJSON generation now allows more control over the properties and wrapping of the output JSON. It may be desireable to inject custom properties into the JSON returned from the query, or even to perform such an injection after the fact. In the former case, the properties are passed into the method and become part of the query. In the latter case, the spatial model foreign key is included in the properties of each feature, allowing the consumer of the JSON to extract and correlate each feature with its corresponding spatial model. The host app can then inject custom properties on a per-feature basis.