Fix bug that allowed invalid geometry to save
We don't ensure the `geog` attribute contains geography before its saved, knowing that it will be coerced when it is saved to the geog column. However, if we have data that would be invalid when in a geographic projection, we need to make sure we change to that projection before validating so those errors are detected.