Change AggregateFeature cache key on refresh
Features are assumed to never be updated, only recreated, which is why the cache key does not check a timestamp, only count and id. Aggregate Features should follow the same rule.

Instead of refreshing the Aggregate Feature, we destroy and create a new one to ensure it gets a different id, and therefore a different cache key.