Optimize tile generation and remove cached KML
This gem is transitioning away from KML to GeoJSON or MVT tile generation because can be handled all client-side. As a result, we no longer have a need for KML cache columns. Instead, we precompute a column of geometry for MVT tiles in SRID 3857 so it doesn't need to be done on-the-fly.