Ensure deterministic SpatialProximity model a/b
When querying for spatial proximities we generally are comparing members of one table against members of another table. However, we don't know which SpatialProximity column the models will end up in, either `model_a` or `model_b`, which means we have to look for both cases.

Instead, we can ensure we know which column each will be stored in by performing a sort at insert time. We can make model a the record with the earliest model name and id (alphanumerically), and therefore determine which columns to query without having to also check for the case where the records were stored in the opposite columns.

The only case that is unaffected is when the comparison models are the same. Though we could also look at the id columns, we're often not specifying a single id, therefore requiring complex SQL in the query instead of a single comparison in Ruby before constructing the query, so we just leave that case unoptimized.

SpatialProximity.normalize can be used to 'flip' the model of rows where the records where not written to the alphanumerically 'correct' column.