Update to using virtual columns
This simplifies the derivative generation step when importing features, and it also allows us to automatically set the feature type without having to determine it in each importer.

This commit also ensures we don't have any more null feature types. These were coming up when features were mult-geometries or geometry collections. Multi-geometries are considered to be their corresponding singular type, and a geometry collection is assumed to be a polygon. The only effect of this assumption is that certain calculations may not optimize by ignoring an aggregate geometry that did not contain geometry of the relevant type.