Don't explode on invalid shapefile parsing
When we load a shapefile with invalid geometry we're exploding. This means we don't get an opportunity to resolve the issue in the make valid steps later.

Instead, we allow unsafe shapefiles to be loaded, in anticipation of those errors either being resolved later or being raised in the shared validity checks.