Fix bug importing KML with altitudes with scientific notation
There are coordinates and altitudes with scientific notiation which end up not matching the Regular Expression we use to sanitize. It appears that we don't actually need to eliminate anything but invalid altitudes as postgis will happily ST_GeomFromKML KML coordinates with altitudes. However, to simply we remove all altitudes so we don't need to distinguish between valid and invalid.